Updated with core developer feedback
